// src/components/character/character.interface.ts
export interface Character {
    name: string;
    height: number,
    mass: number,
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: string[];
    species: string[];
    vehicles: string[];
    starships: string[];
    created: Date | string;
    edited: Date | string;
    url: string;
}